feat: stv questions supports null, blank and invalid votes
All valid and formal ballots is consider in stv  calculus
All the blank and null votes are counted independently.
All invalid vote is counted as a null vote.
If we have n candidates and the answer supports blank and null answers, the results is an array of n+2 elements.
For the element on the index i, considering the first n elements like A, if the candidate i was elected A[i] = 1, otherwise A[i] = 0.