Add election login type enum
The commit adds a new enum field, `election_login_type`, to the
`Election` model. This field represents the type of login allowed for
the election, with options for "Close", "Open", and "Semi Public". This
change also includes modifications to various files to use the new enum
field instead of the previous `private_p` boolean field.