fix: size of the ballot
In ranked votes the size of the ballot is the same as the max number of answer that can be given by each voter.
So, the size of blank and null ballots must be the max number of answers too.
When we count blank and null ballots, we have to make sure about the size of the ballot.