chore: Update tally type for CLOSED question to HOMOMORPHIC
This commit updates the tally type for the CLOSED question to HOMOMORPHIC. It modifies the `questions.py` file in the `app/psifos/model` directory. The `tally_type` attribute of the `AbstractQuestion` class is changed to include the new value. This change is necessary to support the desired functionality of the application.